bpftune: rename tcp_cong_tuner -> tcp_conn_tuner
we will implement alternate per-connection strategy
